Add data sharing engine
Basic engine to perform `data_sharing` with a fediverse server.

It includes all the basic building blocks (APIs + code to fetch
content from fediverse servers authenticating as a proper actor).

Shared content results in the queuing of an (empty) ActiveJob.
FASP need to implement these jobs to do something meaningful.

The debug provider has been extended to use this to subscribe to
content and (so far) fetch new content when it arrives.